About page (#6)
* add about page skeleton

Added about page, named it context to depersonalize it. Added FPO copy.

* add sendmail function

Added sendmail functionality and copy/pastaed a form that needs updating.

* add about content. style contact form

Added styling for content form. Fixed header and footer to represent current state and discourage clicking on links when they are for the current section. Added a dummy 'writing' section just for testing.